feat(python,rust,cli): add support for cumulative functions
